add Jenkinsfile and Dockerfile for corresponding agent
Like with .gitlab-ci.yml the amount of duplication is just hilarious.
Think about the following (minimal) starting point:

I want to build with compilers {clang++,g++} in mode {Debug,Release}.
That makes four combinations which is just not acceptable. There has to
be a way to create that permutations and iterate them without
duplication at all.

In addition, I'd clang++ and g++ to be executed on agents of the same
type but never on the same agent in parallel.

  agent {
    node {
      label 'c++'
    }
  }

With the current implementation, Jenkins creates and assigns random
workspaces making reuse of build artifacts impossible. What is worse,
though is the huge waste of space unless the workspace is removed [3].

[1] https://stackoverflow.com/questions/47521940/how-can-i-force-my-jenkinsfile-parallel-steps-to-use-different-agents
[2] https://jenkins.io/blog/2017/09/25/declarative-1/
[3] https://jenkins.io/doc/pipeline/steps/workflow-basic-steps/#deletedir-recursively-delete-the-current-directory-from-the-workspace